backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
alembic
python-dotenv
pydantic
passlib[bcrypt]
bcrypt==4.0.1
requests==2.32.3   
routeros-api==0.21.0    
python-dotenv==1.0.1
loguru==0.7.2